requirements.txt
gradio
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
geopy
folium
openpyxl
google-generativeai